Revert "Viz: Update export options on widget resize (T562597) (#1803)"
This reverts commit c4f8bc4f4a8380704d4f456842f84044873ca042.
