+++ js/viz/core/helpers.js
@@ -61,50 +61,13 @@ function addChange(settings) {
     buildTotalChanges(proto);
 }
 
-function createChainExecutor() {
-    var chain = [];
-
-    executeChain.add = function(item) { chain.push(item); };
-    return executeChain;
-
-    function executeChain() {
-        var i,
-            ii = chain.length,
-            result;
-        for(i = 0; i < ii; ++i) {
-            result = chain[i].apply(this, arguments);
-        }
-
-        return result;
-    }
-}
-
-function expand(target, name, expander) {
-    var current = target[name];
-    if(current.add) {
-        current.add(expander);
-    } else {
-        current = createChainExecutor();
-        current.add(target[name]);
-        current.add(expander);
-    }
-    target[name] = current;
-}
-
 function addPlugin(plugin) {
-    var proto = this.prototype;
-    proto._plugins.push(plugin);
+    this.prototype._plugins.push(plugin);
     if(plugin.members) {
-        _extend(proto, plugin.members);
+        _extend(this.prototype, plugin.members);
     }
     if(plugin.customize) {
         plugin.customize(this);
-    }
-    if(plugin.extenders) {
-        Object.keys(plugin.extenders).forEach(function(key) {
-            var func = plugin.extenders[key];
-            expand(proto, key, func);
-        }, this);
     }
 }
 
@@ -143,5 +106,3 @@ exports.replaceInherit = function(widget) {
 exports.changes = function() {
     return new Flags();
 };
-
-exports.expand = expand;
