+++ js/viz/tree_map/drilldown.js
@@ -1,7 +1,7 @@
 "use strict";
 
 var proto = require("./tree_map.base").prototype,
-    _expand = require("../core/helpers").expand;
+    _expand = require("./common").expand;
 
 require("./api");
