+++ js/viz/tree_map/hover.js
@@ -5,7 +5,6 @@ var proto = require("./tree_map.base").prototype,
 
     common = require("./common"),
 
-    _expand = require("../core/helpers").expand,
     _parseScalar = require("../core/utils").parseScalar,
     _buildRectAppearance = common.buildRectAppearance,
 
@@ -16,7 +15,7 @@ require("./states");
 
 proto._eventsMap.onHoverChanged = { name: "hoverChanged" };
 
-_expand(proto._handlers, "calculateAdditionalStates", function(states, options) {
+common.expand(proto._handlers, "calculateAdditionalStates", function(states, options) {
     states[1] = options.hoverStyle ? _buildRectAppearance(options.hoverStyle) : {};
 });
 
@@ -37,7 +36,7 @@ require("./tree_map.base").addChange({
 nodeProto.statesMap[1] = 1;
 nodeProto.additionalStates.push(1);
 
-_expand(proto, "_extendProxyType", function(proto) {
+common.expand(proto, "_extendProxyType", function(proto) {
     var that = this;
 
     proto.setHover = function() {
@@ -48,11 +47,11 @@ _expand(proto, "_extendProxyType", function(proto) {
     };
 });
 
-_expand(proto, "_onNodesCreated", function() {
+common.expand(proto, "_onNodesCreated", function() {
     this._hoverIndex = -1;
 });
 
-_expand(proto, "_changeGroupSettings", function() {
+common.expand(proto, "_changeGroupSettings", function() {
     var that = this;
     that._groupHoverEnabled = _parseScalar(that._getOption("group").hoverEnabled, true);
     if(!that._groupHoverEnabled) {
